RATIS-1151. Move the DateStream requests handling code from NettyServerStreamRpc to a new class.
